# perf: make `ape console --help` faster