### What I did

Before:

```
ape console --help  0.84s user 0.12s system 99% cpu 0.968 total
```

After:

```
ape console --help  0.27s user 0.05s system 97% cpu 0.322 total
```

### How I did it

Mostly by avoiding needing to import something from `ape.api.providers`

### How to verify it

```sh
ape console --help
```

### Checklist

<!-- All PRs must complete the following checklist before being merged -->

- [ ] All changes are completed
- [ ] New test cases have been added
- [ ] Documentation has been updated